feat(config): move config.py to app directory
